[Refactor] Move every benchmark-authored input builder into workloads/
Thirty gen_inputs implementations lived in benchmarks/, none of them
reachable by the test stage. Move them to the shared layer the trust model
designates: MoE cases to workloads/moe.py, cumulative to reduction.py, the
elementwise and fp8 generators to elementwise.py, GLA chunkwise to
linear_attention.py, and the pool cases to a new workloads/pool.py.

Correctness oracles and timing baselines do not travel with them. Two
ref_program stubs returned None and are deleted; the real baselines stay
benchmark-side as torch_baseline on a thin subclass of the imported
workload, which is this stage's own content.
