+++ benchmarks/ops/attention/bench_gqa_sliding_window_varlen.py
@@ -22,8 +22,6 @@ _GQA_SLIDING_WINDOW_VARLEN_FWD_BENCH_PARAMS = manifest_params(
     gqa_sliding_window_varlen_args,
     tune=False,
 )
-
-
 def _torch_sliding_window_varlen_fwd(test):
     """Torch SDPA forward baseline: unpack varlen to padded batch, single SDPA call."""
