+++ benchmarks/ops/bench_cumulative.py
@@ -13,32 +13,10 @@ from benchmarks.benchmark_base import (
     ManifestBenchmark,
     workloads_to_params,
 )
-from workloads.workload_base import WorkloadBase
+from workloads.reduction import CumulativeWorkload
 
 _CUMSUM_OP = "CumsumFwdOp"
 _CUMPROD_OP = "CumprodFwdOp"
-
-
-class CumulativeBenchTest(WorkloadBase):
-    def __init__(self, shape: tuple, dtype: torch.dtype, op_kind: str):
-        self.shape = shape
-        self.dtype = dtype
-        self.op_kind = op_kind
-
-    def gen_inputs(self) -> tuple[torch.Tensor]:
-        if self.op_kind == "cumprod":
-            x = torch.rand(*self.shape, dtype=self.dtype, device="cuda") * 0.01 + 0.99
-        else:
-            x = torch.randn(*self.shape, dtype=self.dtype, device="cuda")
-        return (x,)
-
-    def ref_program(self, x: torch.Tensor) -> torch.Tensor:
-        x_f32 = x.float()
-        if self.op_kind == "cumsum":
-            return x_f32.cumsum(dim=-1).to(x.dtype)
-        elif self.op_kind == "cumprod":
-            return x_f32.cumprod(dim=-1).to(x.dtype)
-        raise ValueError(f"Unknown op_kind: {self.op_kind}")
 
 
 def _make_op(shape: tuple, dtype: torch.dtype, op_kind: str):
@@ -53,9 +31,18 @@ def _make_op(shape: tuple, dtype: torch.dtype, op_kind: str):
     return cls(dtype=dtype, dim=-1)
 
 
+class _CumulativeBenchmarkWorkload(CumulativeWorkload):
+    def torch_baseline(self, x: torch.Tensor) -> torch.Tensor:
+        x_f32 = x.float()
+        if self.op_kind == "cumsum":
+            return x_f32.cumsum(dim=-1).to(x.dtype)
+        if self.op_kind == "cumprod":
+            return x_f32.cumprod(dim=-1).to(x.dtype)
+        raise ValueError(f"Unknown op_kind: {self.op_kind}")
+
 @pytest.mark.parametrize("shape, dtype", workloads_to_params(_CUMSUM_OP))
 def test_cumsum_bench(shape: tuple, dtype: torch.dtype) -> None:
-    test = CumulativeBenchTest(shape, dtype, "cumsum")
+    test = _CumulativeBenchmarkWorkload(shape, dtype, "cumsum")
     inputs = test.gen_inputs()
 
     op = _make_op(shape, dtype, "cumsum")
@@ -63,13 +50,13 @@ def test_cumsum_bench(shape: tuple, dtype: torch.dtype) -> None:
     result = bm.profile(op, *inputs)
     BenchmarkReport.record(op, locals(), result, tag="tileops")
 
-    result_bl = bm.profile(test.ref_program, *inputs)
+    result_bl = bm.profile(test.torch_baseline, *inputs)
     BenchmarkReport.record(op, locals(), result_bl, tag="torch")
 
 
 @pytest.mark.parametrize("shape, dtype", workloads_to_params(_CUMPROD_OP))
 def test_cumprod_bench(shape: tuple, dtype: torch.dtype) -> None:
-    test = CumulativeBenchTest(shape, dtype, "cumprod")
+    test = _CumulativeBenchmarkWorkload(shape, dtype, "cumprod")
     inputs = test.gen_inputs()
 
     op = _make_op(shape, dtype, "cumprod")
@@ -77,7 +64,7 @@ def test_cumprod_bench(shape: tuple, dtype: torch.dtype) -> None:
     result = bm.profile(op, *inputs)
     BenchmarkReport.record(op, locals(), result, tag="tileops")
 
-    result_bl = bm.profile(test.ref_program, *inputs)
+    result_bl = bm.profile(test.torch_baseline, *inputs)
     BenchmarkReport.record(op, locals(), result_bl, tag="torch")
