+++ benchmarks/ops/bench_dropout.py
@@ -12,28 +12,18 @@ import torch.nn.functional as F
 
 from benchmarks.benchmark_base import BenchmarkReport, ManifestBenchmark, workloads_to_params
 from tileops.ops.dropout import DropoutOp
-from workloads.workload_base import WorkloadBase
+from workloads.elementwise import DropoutWorkload
 
 _OP_NAME = "DropoutOp"
 
 
-class DropoutBenchCase(WorkloadBase):
-    def __init__(self, shape: tuple, dtype: torch.dtype, p: float = 0.5):
-        self.shape = shape
-        self.n_total = prod(shape)
-        self.dtype = dtype
-        self.p = p
-
-    def gen_inputs(self) -> tuple[torch.Tensor, ...]:
-        return (torch.randn(self.shape, device="cuda", dtype=self.dtype),)
-
-    def ref_program(self, x: torch.Tensor) -> torch.Tensor:
+class _DropoutBenchmarkWorkload(DropoutWorkload):
+    def torch_baseline(self, x: torch.Tensor) -> torch.Tensor:
         return F.dropout(x, p=self.p, training=True)
-
 
 @pytest.mark.parametrize("shape, dtype", workloads_to_params(_OP_NAME))
 def test_dropout_bench(shape: tuple, dtype: torch.dtype) -> None:
-    test = DropoutBenchCase(shape, dtype)
+    test = _DropoutBenchmarkWorkload(shape, dtype)
     (x,) = test.gen_inputs()
 
     op = DropoutOp(p=test.p, seed=42)
@@ -41,7 +31,7 @@ def test_dropout_bench(shape: tuple, dtype: torch.dtype) -> None:
     result = bm.profile(op, x)
     BenchmarkReport.record(op, locals(), result, tag="tileops")
 
-    result_bl = bm.profile(test.ref_program, x)
+    result_bl = bm.profile(test.torch_baseline, x)
     BenchmarkReport.record(op, locals(), result_bl, tag="torch")
