+++ benchmarks/ops/bench_pool.py
@@ -26,6 +26,14 @@ from tileops.ops import (
     MaxPool3dFwdOp,
     MaxPool3dIndicesFwdOp,
 )
+from workloads.pool import (
+    AvgPool1dBenchCase,
+    AvgPool2dBenchCase,
+    AvgPool3dBenchCase,
+    MaxPool1dBenchCase,
+    MaxPool2dBenchCase,
+    MaxPool3dBenchCase,
+)
 
 _AVG_POOL1D_OP_NAME = "AvgPool1dFwdOp"
 _AVG_POOL2D_OP_NAME = "AvgPool2dFwdOp"
@@ -36,44 +44,6 @@ _MAX_POOL2D_OP_NAME = "MaxPool2dFwdOp"
 _MAX_POOL2D_INDICES_OP_NAME = "MaxPool2dIndicesFwdOp"
 _MAX_POOL3D_OP_NAME = "MaxPool3dFwdOp"
 _MAX_POOL3D_INDICES_OP_NAME = "MaxPool3dIndicesFwdOp"
-
-
-class AvgPool1dBenchCase:
-    def __init__(
-        self,
-        n: int,
-        c_in: int,
-        l_in: int,
-        kernel_size: int,
-        stride: Optional[int],
-        padding: int,
-        ceil_mode: bool,
-        count_include_pad: bool,
-        dtype: torch.dtype,
-    ) -> None:
-        self.n = n
-        self.c_in = c_in
-        self.l_in = l_in
-        self.kernel_size = kernel_size
-        self.stride = kernel_size if stride is None else stride
-        self.padding = padding
-        self.ceil_mode = ceil_mode
-        self.count_include_pad = count_include_pad
-        self.dtype = dtype
-
-    def gen_inputs(self) -> tuple[torch.Tensor]:
-        x = torch.randn(self.n, self.c_in, self.l_in, device="cuda", dtype=self.dtype).contiguous()
-        return (x,)
-
-    def ref_program(self, x: torch.Tensor) -> torch.Tensor:
-        return F.avg_pool1d(
-            x,
-            kernel_size=self.kernel_size,
-            stride=self.stride,
-            padding=self.padding,
-            ceil_mode=self.ceil_mode,
-            count_include_pad=self.count_include_pad,
-        )
 
 
 def _avg_pool1d_bench_params() -> list:
@@ -141,51 +111,6 @@ def test_avg_pool1d_bench(
 
     result_bl = bm.profile(test.ref_program, *inputs)
     BenchmarkReport.record(op, locals(), result_bl, tag="torch-ref")
-
-
-class AvgPool2dBenchCase:
-    def __init__(
-        self,
-        n: int,
-        c_in: int,
-        h_in: int,
-        w_in: int,
-        kernel_size: tuple[int, int],
-        stride: Optional[tuple[int, int]],
-        padding: tuple[int, int],
-        ceil_mode: bool,
-        count_include_pad: bool,
-        divisor_override: Optional[int],
-        dtype: torch.dtype,
-    ) -> None:
-        self.n = n
-        self.c_in = c_in
-        self.h_in = h_in
-        self.w_in = w_in
-        self.kernel_size = kernel_size
-        self.stride = kernel_size if stride is None else stride
-        self.padding = padding
-        self.ceil_mode = ceil_mode
-        self.count_include_pad = count_include_pad
-        self.divisor_override = divisor_override
-        self.dtype = dtype
-
-    def gen_inputs(self) -> tuple[torch.Tensor]:
-        x = torch.randn(
-            self.n, self.c_in, self.h_in, self.w_in, device="cuda", dtype=self.dtype
-        ).contiguous()
-        return (x,)
-
-    def ref_program(self, x: torch.Tensor) -> torch.Tensor:
-        return F.avg_pool2d(
-            x,
-            kernel_size=self.kernel_size,
-            stride=self.stride,
-            padding=self.padding,
-            ceil_mode=self.ceil_mode,
-            count_include_pad=self.count_include_pad,
-            divisor_override=self.divisor_override,
-        )
 
 
 def _avg_pool2d_bench_params() -> list:
@@ -271,59 +196,6 @@ def test_avg_pool2d_bench(
 
     result_bl = bm.profile(test.ref_program, *inputs)
     BenchmarkReport.record(op, locals(), result_bl, tag="torch-ref")
-
-
-class AvgPool3dBenchCase:
-    def __init__(
-        self,
-        n: int,
-        c_in: int,
-        d_in: int,
-        h_in: int,
-        w_in: int,
-        kernel_size: tuple[int, int, int],
-        stride: Optional[tuple[int, int, int]],
-        padding: tuple[int, int, int],
-        ceil_mode: bool,
-        count_include_pad: bool,
-        divisor_override: Optional[int],
-        dtype: torch.dtype,
-    ) -> None:
-        self.n = n
-        self.c_in = c_in
-        self.d_in = d_in
-        self.h_in = h_in
-        self.w_in = w_in
-        self.kernel_size = kernel_size
-        self.stride = kernel_size if stride is None else stride
-        self.padding = padding
-        self.ceil_mode = ceil_mode
-        self.count_include_pad = count_include_pad
-        self.divisor_override = divisor_override
-        self.dtype = dtype
-
-    def gen_inputs(self) -> tuple[torch.Tensor]:
-        x = torch.randn(
-            self.n,
-            self.c_in,
-            self.d_in,
-            self.h_in,
-            self.w_in,
-            device="cuda",
-            dtype=self.dtype,
-        ).contiguous()
-        return (x,)
-
-    def ref_program(self, x: torch.Tensor) -> torch.Tensor:
-        return F.avg_pool3d(
-            x,
-            kernel_size=self.kernel_size,
-            stride=self.stride,
-            padding=self.padding,
-            ceil_mode=self.ceil_mode,
-            count_include_pad=self.count_include_pad,
-            divisor_override=self.divisor_override,
-        )
 
 
 def _avg_pool3d_bench_params() -> list:
@@ -412,51 +284,6 @@ def test_avg_pool3d_bench(
 
     result_bl = bm.profile(test.ref_program, *inputs)
     BenchmarkReport.record(op, locals(), result_bl, tag="torch-ref")
-
-
-class MaxPool2dBenchCase:
-    def __init__(
-        self,
-        n: int,
-        c_in: int,
-        h_in: int,
-        w_in: int,
-        kernel_size: tuple[int, int],
-        stride: Optional[tuple[int, int]],
-        padding: tuple[int, int],
-        dilation: tuple[int, int],
-        ceil_mode: bool,
-        dtype: torch.dtype,
-        return_indices: bool = False,
-    ) -> None:
-        self.n = n
-        self.c_in = c_in
-        self.h_in = h_in
-        self.w_in = w_in
-        self.kernel_size = kernel_size
-        self.stride = kernel_size if stride is None else stride
-        self.padding = padding
-        self.dilation = dilation
-        self.ceil_mode = ceil_mode
-        self.dtype = dtype
-        self.return_indices = return_indices
-
-    def gen_inputs(self) -> tuple[torch.Tensor]:
-        x = torch.randn(
-            self.n, self.c_in, self.h_in, self.w_in, device="cuda", dtype=self.dtype
-        ).contiguous()
-        return (x,)
-
-    def ref_program(self, x: torch.Tensor) -> torch.Tensor | tuple[torch.Tensor, torch.Tensor]:
-        return F.max_pool2d(
-            x,
-            kernel_size=self.kernel_size,
-            stride=self.stride,
-            padding=self.padding,
-            dilation=self.dilation,
-            ceil_mode=self.ceil_mode,
-            return_indices=self.return_indices,
-        )
 
 
 def _max_pool2d_bench_params_from_workloads(workloads: list[dict]) -> list:
@@ -595,47 +422,6 @@ def test_max_pool2d_indices_bench(
     BenchmarkReport.record(op, locals(), result_bl, tag="torch-ref")
 
 
-class MaxPool1dBenchCase:
-    def __init__(
-        self,
-        n: int,
-        c_in: int,
-        l_in: int,
-        kernel_size: tuple[int],
-        stride: Optional[tuple[int]],
-        padding: tuple[int],
-        dilation: tuple[int],
-        ceil_mode: bool,
-        dtype: torch.dtype,
-        return_indices: bool = False,
-    ) -> None:
-        self.n = n
-        self.c_in = c_in
-        self.l_in = l_in
-        self.kernel_size = kernel_size
-        self.stride = kernel_size if stride is None else stride
-        self.padding = padding
-        self.dilation = dilation
-        self.ceil_mode = ceil_mode
-        self.dtype = dtype
-        self.return_indices = return_indices
-
-    def gen_inputs(self) -> tuple[torch.Tensor]:
-        x = torch.randn(self.n, self.c_in, self.l_in, device="cuda", dtype=self.dtype).contiguous()
-        return (x,)
-
-    def ref_program(self, x: torch.Tensor) -> torch.Tensor | tuple[torch.Tensor, torch.Tensor]:
-        return F.max_pool1d(
-            x,
-            kernel_size=self.kernel_size,
-            stride=self.stride,
-            padding=self.padding,
-            dilation=self.dilation,
-            ceil_mode=self.ceil_mode,
-            return_indices=self.return_indices,
-        )
-
-
 def _max_pool1d_bench_params_from_workloads(workloads: list[dict]) -> list:
     params = []
     for workload in workloads:
@@ -768,59 +554,6 @@ def test_max_pool1d_indices_bench(
 
     result_bl = bm.profile(test.ref_program, *inputs)
     BenchmarkReport.record(op, locals(), result_bl, tag="torch-ref")
-
-
-class MaxPool3dBenchCase:
-    def __init__(
-        self,
-        n: int,
-        c_in: int,
-        d_in: int,
-        h_in: int,
-        w_in: int,
-        kernel_size: tuple[int, int, int],
-        stride: Optional[tuple[int, int, int]],
-        padding: tuple[int, int, int],
-        dilation: tuple[int, int, int],
-        ceil_mode: bool,
-        dtype: torch.dtype,
-        return_indices: bool = False,
-    ) -> None:
-        self.n = n
-        self.c_in = c_in
-        self.d_in = d_in
-        self.h_in = h_in
-        self.w_in = w_in
-        self.kernel_size = kernel_size
-        self.stride = kernel_size if stride is None else stride
-        self.padding = padding
-        self.dilation = dilation
-        self.ceil_mode = ceil_mode
-        self.dtype = dtype
-        self.return_indices = return_indices
-
-    def gen_inputs(self) -> tuple[torch.Tensor]:
-        x = torch.randn(
-            self.n,
-            self.c_in,
-            self.d_in,
-            self.h_in,
-            self.w_in,
-            device="cuda",
-            dtype=self.dtype,
-        ).contiguous()
-        return (x,)
-
-    def ref_program(self, x: torch.Tensor) -> torch.Tensor | tuple[torch.Tensor, torch.Tensor]:
-        return F.max_pool3d(
-            x,
-            kernel_size=self.kernel_size,
-            stride=self.stride,
-            padding=self.padding,
-            dilation=self.dilation,
-            ceil_mode=self.ceil_mode,
-            return_indices=self.return_indices,
-        )
 
 
 def _max_pool3d_bench_params_from_workloads(workloads: list[dict]) -> list:
