+++ workloads/moe.py
@@ -1,3 +1,5 @@
+from typing import Any
+
 import torch
 
 from workloads.workload_base import WorkloadBase
@@ -141,3 +143,195 @@ class MoeGroupedGemmNopadWorkload(WorkloadBase):
         a = torch.randn(self.numel, self.k, dtype=self.dtype, device=dev) * 0.02
         b = torch.randn(self.num_experts, self.n, self.k, dtype=self.dtype, device=dev) * 0.02
         return a, b, true_sizes, true_offsets
+
+
+class FusedMoeWorkload(WorkloadBase):
+    """Inputs for a single FusedMoe benchmark configuration."""
+
+    def __init__(
+        self,
+        num_tokens: int,
+        num_experts: int,
+        top_k: int,
+        hidden_size: int,
+        ffn_size: int,
+        scoring_func: str,
+        renormalize: bool,
+        with_correction_bias: bool,
+        routed_scaling_factor: float,
+        dtype: torch.dtype,
+    ):
+        self.num_tokens = num_tokens
+        self.num_experts = num_experts
+        self.top_k = top_k
+        self.hidden_size = hidden_size
+        self.ffn_size = ffn_size
+        self.scoring_func = scoring_func
+        self.renormalize = renormalize
+        self.with_correction_bias = with_correction_bias
+        self.routed_scaling_factor = routed_scaling_factor
+        self.dtype = dtype
+
+    def gen_inputs(self):
+        torch.manual_seed(42)
+        dev = "cuda"
+        hidden = torch.randn(
+            self.num_tokens, self.hidden_size, dtype=self.dtype, device=dev,
+        )
+        gating = torch.randn(
+            self.num_tokens, self.num_experts, dtype=torch.float32, device=dev,
+        )
+        correction_bias = (
+            torch.randn(self.num_experts, dtype=torch.float32, device=dev) * 0.1
+            if self.with_correction_bias else None
+        )
+        w_gate_up = torch.randn(
+            self.num_experts, self.ffn_size * 2, self.hidden_size,
+            dtype=self.dtype, device=dev,
+        ) * 0.02
+        w_down = torch.randn(
+            self.num_experts, self.hidden_size, self.ffn_size,
+            dtype=self.dtype, device=dev,
+        ) * 0.02
+        return hidden, gating, correction_bias, w_gate_up, w_down
+
+
+class SharedFusedMoeWorkload(WorkloadBase):
+    def __init__(
+        self,
+        num_tokens,
+        num_experts,
+        top_k,
+        hidden_size,
+        ffn_size,
+        shared_ffn_size,
+        scoring_func,
+        renormalize,
+        with_correction_bias,
+        routed_scaling_factor,
+        dtype,
+    ):
+        self.num_tokens = num_tokens
+        self.num_experts = num_experts
+        self.top_k = top_k
+        self.hidden_size = hidden_size
+        self.ffn_size = ffn_size
+        self.shared_ffn_size = shared_ffn_size
+        self.scoring_func = scoring_func
+        self.renormalize = renormalize
+        self.with_correction_bias = with_correction_bias
+        self.routed_scaling_factor = routed_scaling_factor
+        self.dtype = dtype
+
+    def gen_inputs(self):
+        torch.manual_seed(42)
+        dev = "cuda"
+        hidden = torch.randn(
+            self.num_tokens, self.hidden_size, dtype=self.dtype, device=dev
+        )
+        gating = torch.randn(
+            self.num_tokens, self.num_experts, dtype=self.dtype, device=dev
+        )
+        correction_bias = (
+            torch.randn(self.num_experts, dtype=torch.float32, device=dev) * 0.1
+            if self.with_correction_bias else None
+        )
+        w_gate_up = torch.randn(
+            self.num_experts, self.ffn_size * 2, self.hidden_size,
+            dtype=self.dtype, device=dev,
+        ) * 0.02
+        w_down = torch.randn(
+            self.num_experts, self.hidden_size, self.ffn_size,
+            dtype=self.dtype, device=dev,
+        ) * 0.02
+        # Shared expert weights: gate+up concatenated [2*Fs, H], down [H, Fs]
+        shared_w_gate_up = torch.randn(
+            self.shared_ffn_size * 2, self.hidden_size, dtype=self.dtype, device=dev
+        ) * 0.02
+        shared_w_down = torch.randn(
+            self.hidden_size, self.shared_ffn_size, dtype=self.dtype, device=dev
+        ) * 0.02
+        return hidden, gating, correction_bias, w_gate_up, w_down, shared_w_gate_up, shared_w_down
+
+
+class MoeExpertsWorkload(WorkloadBase):
+    def __init__(self, num_tokens, num_experts, top_k, hidden_size, ffn_size, dtype):
+        self.num_tokens = num_tokens
+        self.num_experts = num_experts
+        self.top_k = top_k
+        self.hidden_size = hidden_size
+        self.ffn_size = ffn_size
+        self.dtype = dtype
+
+    def gen_inputs(self):
+        torch.manual_seed(42)
+        dev = "cuda"
+        hidden = torch.randn(self.num_tokens, self.hidden_size, dtype=self.dtype, device=dev)
+        w1 = torch.randn(self.num_experts, self.ffn_size * 2, self.hidden_size, dtype=self.dtype, device=dev) * 0.02
+        w2 = torch.randn(self.num_experts, self.hidden_size, self.ffn_size, dtype=self.dtype, device=dev) * 0.02
+        topk_weights = torch.softmax(
+            torch.randn(self.num_tokens, self.top_k, dtype=torch.float32, device=dev), dim=-1
+        )
+        topk_ids = torch.randint(0, self.num_experts, (self.num_tokens, self.top_k), dtype=torch.int32, device=dev)
+        return hidden, w1, w2, topk_weights, topk_ids
+
+
+class MoeFusedActivationWorkload(WorkloadBase):
+    """Workload descriptor for fused vs unfused activation benchmark."""
+
+    def __init__(
+        self,
+        num_tokens: int,
+        hidden_size: int,
+        ffn_size: int,
+        num_experts: int,
+        top_k: int,
+        dtype: torch.dtype,
+    ):
+        self.num_tokens = num_tokens
+        self.hidden_size = hidden_size
+        self.ffn_size = ffn_size
+        self.num_experts = num_experts
+        self.top_k = top_k
+        self.dtype = dtype
+        # Primary shape: (num_tokens, hidden_size) — the token tensor footprint.
+        self.shape: tuple[int, int] = (num_tokens, hidden_size)
+
+    def gen_inputs(self) -> tuple[Any, ...]:
+        torch.manual_seed(42)
+        dev = "cuda"
+        hidden = torch.randn(
+            self.num_tokens, self.hidden_size, dtype=self.dtype, device=dev,
+        )
+        w_gate_up = torch.randn(
+            self.num_experts, self.ffn_size * 2, self.hidden_size,
+            dtype=self.dtype, device=dev,
+        ) * 0.02
+        w_down = torch.randn(
+            self.num_experts, self.hidden_size, self.ffn_size,
+            dtype=self.dtype, device=dev,
+        ) * 0.02
+        topk_weights = torch.softmax(
+            torch.randn(self.num_tokens, self.top_k, dtype=torch.float32, device=dev),
+            dim=-1,
+        )
+        topk_ids = torch.randint(
+            0, self.num_experts,
+            (self.num_tokens, self.top_k), dtype=torch.int32, device=dev,
+        )
+        return hidden, w_gate_up, w_down, topk_weights, topk_ids
+
+
+class MoeSharedExpertMlpWorkload(WorkloadBase):
+    def __init__(self, num_tokens, hidden_size, ffn_size, dtype):
+        self.num_tokens = num_tokens
+        self.hidden_size = hidden_size
+        self.ffn_size = ffn_size
+        self.dtype = dtype
+
+    def gen_inputs(self):
+        device = torch.device("cuda")
+        hidden = torch.randn(self.num_tokens, self.hidden_size, dtype=self.dtype, device=device)
+        w_gate_up = torch.randn(self.ffn_size * 2, self.hidden_size, dtype=self.dtype, device=device)
+        w_down = torch.randn(self.hidden_size, self.ffn_size, dtype=self.dtype, device=device)
+        return hidden, w_gate_up, w_down
